requirements.txt
requests
translate
python-dotenv
openai